+++ torch/ao/nn/quantized/_reference/__init__.py
@@ -1,0 +1,1 @@
+from .modules import *  # noqa: F403
