+++ torch/ao/quantization/backend_config/tensorrt.py
@@ -42,7 +42,7 @@ def get_tensorrt_backend_config_dict():
             weighted_op_qint8_dtype_config,
         ],
         "root_module": torch.nn.Linear,
-        "reference_quantized_module_for_root": torch.nn.quantized._reference.Linear,
+        "reference_quantized_module_for_root": torch.ao.nn.quantized._reference.Linear,
     }
     # TODO: maybe make "pattern" to be a list of patterns
     # TODO: current patterns are the ones after fusion, we will want to expose fusion
@@ -73,7 +73,7 @@ def get_tensorrt_backend_config_dict():
             weighted_op_qint8_dtype_config,
         ],
         "root_module": torch.nn.Linear,
-        "reference_quantized_module_for_root": torch.nn.quantized._reference.Linear,
+        "reference_quantized_module_for_root": torch.ao.nn.quantized._reference.Linear,
         "qat_module": nniqat.LinearReLU,
     }
     linear_relu_qat_config = {
@@ -83,7 +83,7 @@ def get_tensorrt_backend_config_dict():
             weighted_op_qint8_dtype_config,
         ],
         "root_module": torch.nn.Linear,
-        "reference_quantized_module_for_root": torch.nn.quantized._reference.Linear,
+        "reference_quantized_module_for_root": torch.ao.nn.quantized._reference.Linear,
     }
     conv_module_config = {
         "pattern": torch.nn.Conv2d,
@@ -92,7 +92,7 @@ def get_tensorrt_backend_config_dict():
             weighted_op_qint8_dtype_config,
         ],
         "root_module": torch.nn.Conv2d,
-        "reference_quantized_module_for_root": torch.nn.quantized._reference.Conv2d,
+        "reference_quantized_module_for_root": torch.ao.nn.quantized._reference.Conv2d,
         "qat_module": nnqat.Conv2d,
     }
     conv_qat_config = {
@@ -102,7 +102,7 @@ def get_tensorrt_backend_config_dict():
             weighted_op_qint8_dtype_config,
         ],
         "root_module": torch.nn.Conv2d,
-        "reference_quantized_module_for_root": torch.nn.quantized._reference.Conv2d,
+        "reference_quantized_module_for_root": torch.ao.nn.quantized._reference.Conv2d,
     }
     conv1d_relu_fused_config = {
         "pattern": nni.ConvReLU1d,
@@ -111,7 +111,7 @@ def get_tensorrt_backend_config_dict():
             weighted_op_qint8_dtype_config,
         ],
         "root_module": torch.nn.Conv1d,
-        "reference_quantized_module_for_root": torch.nn.quantized._reference.Conv1d,
+        "reference_quantized_module_for_root": torch.ao.nn.quantized._reference.Conv1d,
     }
     conv2d_relu_fused_config = {
         "pattern": nni.ConvReLU2d,
@@ -120,7 +120,7 @@ def get_tensorrt_backend_config_dict():
             weighted_op_qint8_dtype_config,
         ],
         "root_module": torch.nn.Conv2d,
-        "reference_quantized_module_for_root": torch.nn.quantized._reference.Conv2d,
+        "reference_quantized_module_for_root": torch.ao.nn.quantized._reference.Conv2d,
         "qat_module": nniqat.ConvReLU2d,
     }
     conv2d_relu_qat_config = {
@@ -130,7 +130,7 @@ def get_tensorrt_backend_config_dict():
             weighted_op_qint8_dtype_config,
         ],
         "root_module": torch.nn.Conv2d,
-        "reference_quantized_module_for_root": torch.nn.quantized._reference.Conv2d,
+        "reference_quantized_module_for_root": torch.ao.nn.quantized._reference.Conv2d,
     }
     conv3d_relu_fused_config = {
         "pattern": nni.ConvReLU3d,
@@ -139,7 +139,7 @@ def get_tensorrt_backend_config_dict():
             weighted_op_qint8_dtype_config,
         ],
         "root_module": torch.nn.Conv3d,
-        "reference_quantized_module_for_root": torch.nn.quantized._reference.Conv3d,
+        "reference_quantized_module_for_root": torch.ao.nn.quantized._reference.Conv3d,
         "qat_module": nniqat.ConvReLU3d,
     }
     conv2d_relu_mf_config = {
