+++ torch/ao/quantization/quantization_mappings.py
@@ -9,7 +9,7 @@ import torch.nn.intrinsic.quantized as nniq
 import torch.nn.intrinsic.quantized.dynamic as nniqd
 import torch.nn.intrinsic.qat as nniqat
 import torch.ao.nn.quantized as nnq
-import torch.nn.quantized._reference as nnqr
+import torch.ao.nn.quantized._reference as nnqr
 import torch.ao.nn.quantized.dynamic as nnqd
 import torch.nn.qat as nnqat
 import torch.nn.qat.dynamic as nnqatd
